+++ tmtccmd/cfdp/mib.py
@@ -33,7 +33,16 @@ class DefaultFaultHandlerBase(ABC):
 
     For each error reported by :py:meth:`report_fault`, the appropriate fault handler callback
     will be called. The user provides the callbacks by providing a custom class which implements
-    these base class and all abstract fault handler callbacks.
+    this base class and all abstract fault handler callbacks.
+
+    Some note on the provided default settings:
+
+    - Checksum failures will be ignored by default. This is because for unacknowledged transfers,
+      cancelling the transfer immediately would interfere with the check limit mechanism specified
+      in chapter 4.6.3.3.
+    - Unsupported checksum types will also be ignored by default. Even if the checksum type is
+      not supported the file transfer might still have worked properly.
+
     """
 
     def __init__(self):
@@ -43,26 +52,44 @@ class DefaultFaultHandlerBase(ABC):
             ConditionCode.POSITIVE_ACK_LIMIT_REACHED: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.KEEP_ALIVE_LIMIT_REACHED: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.INVALID_TRANSMISSION_MODE: FaultHandlerCode.NOTICE_OF_CANCELLATION,
-            ConditionCode.FILE_CHECKSUM_FAILURE: FaultHandlerCode.NOTICE_OF_CANCELLATION,
+            ConditionCode.FILE_CHECKSUM_FAILURE: FaultHandlerCode.IGNORE_ERROR,
             ConditionCode.FILE_SIZE_ERROR: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.FILESTORE_REJECTION: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.NAK_LIMIT_REACHED: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.INACTIVITY_DETECTED: FaultHandlerCode.NOTICE_OF_CANCELLATION,
             ConditionCode.CHECK_LIMIT_REACHED: FaultHandlerCode.NOTICE_OF_CANCELLATION,
-            ConditionCode.UNSUPPORTED_CHECKSUM_TYPE: FaultHandlerCode.NOTICE_OF_CANCELLATION,
+            ConditionCode.UNSUPPORTED_CHECKSUM_TYPE: FaultHandlerCode.IGNORE_ERROR,
         }
 
     def get_fault_handler(self, condition: ConditionCode) -> Optional[FaultHandlerCode]:
         return self._handler_dict.get(condition)
 
     def set_handler(self, condition: ConditionCode, handler: FaultHandlerCode):
+        """
+        Raises
+        -------
+
+        ValueError
+            Invalid condition code which is not applicable for fault handling procedures.
+        """
         if condition not in self._handler_dict:
-            return
+            raise ValueError(
+                f"condition code {condition!r} not applicable for fault handling procedures"
+            )
         self._handler_dict.update({condition: handler})
 
     def report_fault(self, condition: ConditionCode):
+        """
+        Raises
+        -------
+
+        ValueError
+            Invalid condition code which is not applicable for fault handling procedures.
+        """
         if condition not in self._handler_dict:
-            return
+            raise ValueError(
+                f"condition code {condition!r} not applicable for fault handling procedures"
+            )
         fh_code = self._handler_dict.get(condition)
         if fh_code == FaultHandlerCode.NOTICE_OF_CANCELLATION:
             self.notice_of_cancellation_cb(condition)
@@ -95,9 +122,9 @@ class EntityType(enum.IntEnum):
     RECEIVING = 1
 
 
-class CheckLimitProvider(ABC):
+class CheckTimerProvider(ABC):
     @abc.abstractmethod
-    def provide_check_limit(
+    def provide_check_timer(
         self,
         local_entity_id: UnsignedByteField,
         remote_entity_id: UnsignedByteField,
@@ -157,11 +184,15 @@ class RemoteEntityCfg:
     default_transmission_mode:
         If the transmission mode is not supplied as part of the Put Request, it will be
         determined from this field in the remote configuration.
+    disposition_on_cancellation:
+        Determines whether an incomplete received file is discard on transaction cancellation.
+        Defaults to False.
     crc_type:
         Default checksum type used to calculate for all file transmissions to this remote entity.
-    check_limit_provider:
-        Both the source and destination handler use a check limit for the unacknowledged mode.
-        This generic provider allows the user to configure the check limit at run time as well.
+    check_limit:
+        this timer determines the expiry period for incrementing a check counter after an EOF PDU
+        is received for an incomplete file transfer. This allows out-of-order reception of file
+        data PDUs and EOF PDUs. Also see 4.6.3.3 of the CFDP standard. Defaults to 2.
 
     """
 
@@ -172,7 +203,8 @@ class RemoteEntityCfg:
     crc_on_transmission: bool
     default_transmission_mode: TransmissionMode
     crc_type: ChecksumType
-    check_limit_provider: Optional[CheckLimitProvider]
+    check_limit: int = 2
+    disposition_on_cancellation: bool = False
     # NOTE: Only this version is supported
     cfdp_version: int = CFDP_VERSION_2